+++ lib/Algo/Methods/InstantRunoff/InstantRunoff.php
@@ -119,7 +119,7 @@ class InstantRunoff extends Method implements MethodInterface
 
             $weight = $oneVote->getWeight($election);
 
-            foreach ($oneVote->getContextualRanking($election, false) as $oneRank) :
+            foreach ($oneVote->getContextualRankingWithoutSort($election) as $oneRank) :
                 foreach ($oneRank as $oneCandidate) :
                     if (\count($oneRank) !== 1) :
                         break;