+++ lib/Algo/Methods/STV/SingleTransferableVote.php
@@ -116,7 +116,7 @@ class SingleTransferableVote extends Method implements MethodInterface
             $LoserBonusWeight = 0;
 
             $firstRank = true;
-            foreach ($oneVote->getContextualRanking($election, false) as $oneRank) :
+            foreach ($oneVote->getContextualRankingWithoutSort($election) as $oneRank) :
                 foreach ($oneRank as $oneCandidate) :
                     if (\count($oneRank) !== 1): break; endif;